-Wsign-compare

+++ include/network.h
@@ -177,7 +177,7 @@ private:
     }
 
     const vec_t& forward_propagation(const vec_t& in) {
-        if (in.size() != in_dim())
+        if (in.size() != (size_t)in_dim())
             throw nn_error("input dimension mismatch");
         return layers_.head()->forward_propagation(in);
     }
